# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraApi",
        "description": "LaraApi project API documents",
        "contact": {
            "email": "russell57260620@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tag": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "index",
                "description": "Display a listing of the tags table.",
                "operationId": "getTagIndex",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tag"
                ],
                "summary": "store",
                "description": "Store a newly created tag in tags table.",
                "operationId": "postTagStore",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "head": {
                "tags": [
                    "Tag"
                ],
                "summary": "index",
                "description": "Display a listing of the tags table.",
                "operationId": "headTagIndex",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            }
        },
        "/api/tag/{id}": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "show",
                "description": "Display the specified tag with tag id.",
                "operationId": "getTagShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tag"
                ],
                "summary": "update",
                "description": "Update the specified tag in tags table.",
                "operationId": "putTagUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tag"
                ],
                "summary": "destroy",
                "description": "Remove the specified tag from tags table.",
                "operationId": "deleteTagDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "head": {
                "tags": [
                    "Tag"
                ],
                "summary": "show",
                "description": "Display the specified tag with tag id.",
                "operationId": "headTagShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tag"
                ],
                "summary": "update",
                "description": "Update the specified tag in tags table.",
                "operationId": "patchTagUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    },
...
}